Refactor how layer visibility is handled to improve map performance and rename useSelectedLayer hook accordingly
