--- src/hooks/useSelectedLayer/index.ts
@@ -1,1 +1,0 @@
-export * from "./useSelectedLayer";
